renamed folder api to apollo
